feat: Add "start" script to package.json
This commit adds a new "start" script to the package.json file. The "start" script runs the server using the compiled code in the dist folder. This allows for easy starting of the server without needing to manually run the compiled file. The script is configured to run the "node dist/server.mjs" command. This change improves the development workflow by providing a convenient way to start the server.

Closes #456